[#73] Split utils module
- Context, JsonConfig, and *Path get own modules
- update Context class and add session placeholder
