+++ utils/__init__.py
@@ -1,5 +1,8 @@
 """iBridges utility modules
 """
 
+from . import context
+from . import json_config
 from . import elabConnector
+from . import path
 from . import utils
